requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
av==13.1.0
bitsandbytes==0.41.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
decord==0.6.0
einops==0.6.1
einops-exts==0.0.4
executing==2.1.0
fastapi==0.115.5
ffmpeg==1.4
ffmpy==0.4.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
fvcore==0.1.5.post20221221
google-generativeai==0.8.5
gradio==5.7.0
gradio_client==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
iopath==0.1.10
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
latex2mathml==3.77.0
linkify-it-py==2.0.3
lit==18.1.8
llvmlite==0.43.0
markdown-it-py==2.2.0
markdown2==2.5.1
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==10.5.0
moviepy==2.1.1
mpmath==1.3.0
multidict==6.1.0
narwhals==1.14.2
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper @ git+https://github.com/openai/whisper.git@fc5ded7d9045c693692f13853857c3f8baea3a7b
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
peft==0.4.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.0.0
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.8.0
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==0.6.13
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.31.0
triton==2.0.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.32.1
-e git+https://github.com/PKU-YuanGroup/Video-LLaVA@108940f06fa80e7e71886f065452b69120eead85#egg=videollava
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.13
yacs==0.1.8
yarl==1.18.0